% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_so_on.R
\name{and_so_on}
\alias{and_so_on}
\title{Formats tables to show top and bottom rows.}
\usage{
and_so_on(data, top = 3, bottom = 2, message = "... and so on ...")
}
\arguments{
\item{data}{A data frame to display}

\item{top}{An integer. How many of the first rows to show.}

\item{bottom}{Like \code{top}, for for the bottom rows.}

\item{message}{Character string to place between the top and bottom rows.}
}
\description{
Formats tables to show top and bottom rows.
}
