% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askMC.R
\name{askMC}
\alias{askMC}
\alias{askGS}
\alias{fix_dollar_signs}
\alias{format_latex_answers}
\alias{letter_counter}
\title{Formatting multiple-choice questions for the book}
\usage{
askMC(
  prompt = "The question prompt",
  ...,
  id = NULL,
  right_one = NULL,
  inline = FALSE,
  random_answer_order = FALSE,
  allow_retry = TRUE,
  correct = "Right!",
  incorrect = "Sorry.",
  message = NULL,
  post_message = NULL,
  submit_button = "Check answer",
  try_again_button = "Try again",
  allow_multiple_correct = FALSE,
  show_feedback = TRUE,
  out_format = c("PDF", "Markdown", "GradeScope"),
  item_label = "Part ",
  show_answers = FALSE
)

askGS(...)

fix_dollar_signs(str)

format_latex_answers(AT, linechars = 50)

letter_counter()
}
\description{
Formatting multiple-choice questions for the book
}
