% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dereference-hash.R
\name{dereference_hash}
\alias{dereference_hash}
\title{Dereference a hash code into a link and a number.}
\usage{
dereference_hash(hash)
}
\arguments{
\item{hash}{The alphanumeric character string identifying the object.}
}
\description{
Dereference a hash code into a link and a number.
}
