% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset-exercises.R
\name{format_exercises}
\alias{format_exercises}
\title{Utilities for laying out exercises}
\usage{
format_exercises(
  Roster,
  label_format = c("numeric", "alpha", "ALPHA", "roman", "ROMAN"),
  template = "## Exer. {number}\\n\\n **Exercise {number}**\\n\\n{contents}\\n\\n"
)
}
\arguments{
\item{Roster}{A data frame with a \code{fname} column
containing the file name of the exercise}

\item{chap}{A character string identifying the chapter in the exercise label}

\item{start_num}{Integer > 0 or letter "a" to "z" specifying where to start in the label sequence}
}
\description{
Utilities for laying out exercises
}
