% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPaper.R
\name{graphPaper}
\alias{graphPaper}
\alias{graph_paper}
\title{Traditional graph-paper axes}
\usage{
graph_paper(
  P = NULL,
  xticks = 0:5,
  yticks = xticks,
  show_labels = FALSE,
  no_axes = TRUE,
  xlabels = xticks,
  ylabels = yticks,
  extend = 0.1,
  xlim = c(min(xticks), max(xticks)),
  ylim = c(min(yticks), max(yticks)),
  xlab = "x",
  ylab = "y",
  maxxlabels = 7,
  maxylabels = 7,
  ...
)
}
\arguments{
\item{xticks}{numerical vector listing the position of x-axis ticks}

\item{yticks}{like \code{xticks} but for y-axis}

\item{show_labels}{logical (default: FALSE) whether to place axis labels.}

\item{no_axes}{logical (default: TRUE) Do not display the x and y axes}

\item{xlabels}{as in \code{plot}, sets the y-axis label}

\item{ylabels}{similarly, set the y-axis label}

\item{xlab}{Label for horizontal axis}

\item{ylab}{Label for vertical axis}
}
\value{
a ggplot2 object
}
\description{
\code{graphPaper} plots out traditional graph-paper axes for making
mathematical plots.  Whereas standard R graphics put the axes at the
edge of the plotting window, \code{graphPaper} puts them through the zeros
and lets you set explicitly the location of tick marks and graph-paper rules.
You can then plot over the axes by using functions such as \code{lines},
\code{points}, or \code{funPlot} (with \code{add=TRUE}).
}
\examples{
graph_paper(xticks=-5:5, yticks=seq(-1, 1, by=.25),
            ylabels = c(-1, 1)) \%>\%
  graphFun(sin((x^2) / 3)~x, x=range(-4, 4), color="blue")

}
\author{
Daniel Kaplan (\email{kaplan@macalester.edu})
}
