% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_items.R
\name{latex_vec}
\alias{latex_vec}
\title{Construct latex markup for vectors}
\usage{
latex_vec(n = 3, name = "v", digits = 3, seed = NULL, v = NULL)
}
\arguments{
\item{n}{How many components}

\item{digits}{Number of digits for integer vector}

\item{v}{and alternative way to specify the numbers in the vectors}
}
\description{
Construct latex markup for vectors
}
