% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\alias{exercise_navpoint}
\title{Places position marks in bookdown file}
\usage{
mark(id)

exercise_navpoint(num, perm_id, fname = "no file specified")
}
\arguments{
\item{id}{character or numeric. Id for the milestone. For proper functionality
this should be unique within the document.}
}
\description{
Add a milestone markers to the text of the bookdown document
making it easier to readers to refer to a particular place in the
document.
}
\examples{
\dontrun{
`r mark(4510)`  # place in Rmd file
}

}
