% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocks.R
\name{matrix_block}
\alias{matrix_block}
\title{Variations on RMarkdown blocks}
\usage{
matrix_block(
  cmd,
  digits = 4,
  ignore_cmd = FALSE,
  phrase = "      ",
  display = FALSE,
  inline = TRUE,
  width = "40\%"
)
}
\arguments{
\item{cmd}{The expression to be run}

\item{digits}{How many digits to round to}

\item{ignore_cmd}{If TRUE, just typeset the result}

\item{phrase}{Character string separating code from output}

\item{display}{If TRUE, use $$ rather than single $}

\item{inline}{If TRUE, typeset the output next to the command}

\item{width}{Character string such as "40\%" saying how big the command
block should be.}
}
\description{
Creates from an inline chunk output that resembles a regular
RMarkdown chunk, but with more compact formatting.
}
