% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askMC.R
\name{MC_simple_format}
\alias{MC_simple_format}
\title{Formatting multiple-choice questions for the book}
\usage{
MC_simple_format(
  prompt = "The question prompt",
  ...,
  item_label = "Part",
  out_format = c("Markdown", "PDF")
)
}
\description{
Formatting multiple-choice questions for the book
}
