% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2html.R
\name{md2html}
\alias{md2html}
\title{Prepare markdown-like syntax for display in Shiny}
\usage{
md2html(s)
}
\arguments{
\item{s}{character string containing simple markdown/latex-math}
}
\value{
Another character string containing the equivalent of s in HTML.
This is suitable for an htmlOutput() display in Shiny. You'll need to
escape the HTML by using the html() function in your shiny::renderText()
}
\description{
Prepare markdown-like syntax for display in Shiny
}
