% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-axes.R
\name{scale_shift}
\alias{scale_shift}
\alias{plot_scaled_input}
\title{Draw parallel axes with a linear transform}
\usage{
scale_shift(min, max, r, x0 = 0, nticks = 10, color = "blue")

plot_scaled_input(ftilde, domain, r, x0, nticks = 10, color = "blue")
}
\arguments{
\item{min}{lower end of domain}

\item{max}{upper end of domain}

\item{r}{r parameter in linear scaling \code{r*(x-x0)}}

\item{x0}{x0 parameter in linear scaling \code{r*(x-x0)}}

\item{nticks}{number of ticks on new axis (add_scale)}

\item{color}{color of new axis (e.g. "blue")}

\item{ftilde}{tilde expression as in \code{slice_plot()}}

\item{domain}{domain specifier as in \code{slice_plot()}}
}
\value{
a ggplot2 object
}
\description{
Draws two parallel axes. The lower axis shows the input to
the function without a scaled input, e.g. the input to the pattern book
function. The upper axis shows the input that goes into the scaling
before being handed off to the pattern-book function.
}
\details{
\code{plot_scaled_input()} is for drawing pattern-book functions on
their natural unscaled input (plain "x"), but then adds a second scale
that shows the input to the scaled function.

Only handles linear transformations: r*(x-x0)
}
\examples{
scale_shift(-40, 100, 9/5, 32, 10, color="magenta")
plot_scaled_input(exp(x) ~ x, domain(x=c(-2,2)), 2.4, 1)

}
