% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase-line.R
\name{phase_line}
\alias{phase_line}
\title{Draw a one-dimensional phase space with arrows.}
\usage{
phase_line(tilde, domain, narrows = 15, nix_dyn = FALSE, transform = I)
}
\arguments{
\item{tilde}{A tilde expression giving the dynamical function}

\item{domain}{The domain to show of the graphics, as in \code{slice_plot()}.}

\item{narrows}{How many arrows to place on the phase line}

\item{nix_dyn}{If \code{TRUE}, don't show the dynamical function.}

\item{transform}{A function to use to make the short arrows
long enough to stand out. Try \code{transform=sqrt}.}
}
\description{
Draw a one-dimensional phase space with arrows.
}
\examples{
phase_line(4*x*(1-x) ~ x, domain(x=0:1), narrows=10)

}
