% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riemann_bars.R
\name{Riemann_bars}
\alias{Riemann_bars}
\title{Create dataframe for graphing Riemann bars}
\usage{
Riemann_bars(tilde, h, domain, ...)
}
\arguments{
\item{tilde}{tilde expression defining a function}

\item{h}{width of bars}

\item{domain}{as in slice_plot()}

\item{...}{parameters for tilde expression}
}
\description{
Create dataframe for graphing Riemann bars
}
\examples{
Pts <- Riemann_bars(x*dnorm(x) ~ x, h=0.1, domain(x=c(-3,3)))
\dontrun{
gf_rect(ymin + ymax ~ xmin + xmax, data = Pts,
fill=~ color, alpha=0.3, color="black") \%>\%
gf_refine(scale_fill_identity())
}

}
