% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize.R
\name{segmentize}
\alias{segmentize}
\title{Break a function into segments, giving offset and relative values}
\usage{
segmentize(tilde, h, domain, npts = 10, ...)
}
\description{
Break a function into segments, giving offset and relative values
}
\examples{
Pts <- segmentize(x*dnorm(x) ~ x, .1, list(x=c(-3,3)))
\dontrun{
gf_segment(ymin + ymax ~ xmin + xmax, data = Pts)
}
}
