% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_abcd.R
\name{show_abcd}
\alias{show_abcd}
\title{Draw ABCD linear dynamics phase plane flow}
\usage{
show_abcd(
  a,
  b,
  c,
  d,
  which = c("both", "x", "y"),
  show_sign = FALSE,
  pow = 0.3
)
}
\arguments{
\item{a, }{b, c, d: Four components of the dynamics matrix.}

\item{which}{Which variable(s) to show the positive side of the nullcline in color}

\item{show_sign}{When \code{TRUE}, show the flow vectors all the same length}

\item{pow}{Exponent to use for display length versus actual length of a flow vector}
}
\description{
Draw ABCD linear dynamics phase plane flow
}
\details{
\code{pow} can be used to make the very short flow vectors near the origin look
bigger. Default is 0.3.
}
\examples{
Znotes::show_abcd(-.5, -1, 1, .7)

}
