% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_exercise_list.R
\name{skills_exercise_list}
\alias{skills_exercise_list}
\title{Format exercise-by-exercise skill list}
\usage{
skills_exercise_list(fname = "Skill_list.csv")
}
\description{
Format exercise-by-exercise skill list
}
