% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success.R
\name{random_regret}
\alias{random_regret}
\alias{random_success}
\title{Generates a random success or regret message}
\usage{
random_regret()

random_success()
}
\description{
Just for a bit of variety in multiple-choice feedback.
}
