% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zdata.R
\name{Zdata}
\alias{Zdata}
\title{Read a data file from the Znotes repository}
\usage{
Zdata(dname, ...)
}
\arguments{
\item{dname}{the complete name of the .csv file, e.g. \code{"USAFA/hawaii.csv"}}

\item{...}{other arguments to readr::read_csv()}
}
\description{
The \code{{Znotes}} repository on GitHub has an associated web site. Authorized
instructors can add CSV files to that site. These data files can then
be used in the Sandbox or in the R console.
}
\examples{
hawaii <- Zdata("USAFA/hawaii.csv")

}
