% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_so_on.R
\name{and_so_on}
\alias{and_so_on}
\title{Formats tables to show top and bottom rows.}
\usage{
and_so_on(
  data,
  top = 3,
  bottom = 2,
  message = ifelse(bottom == 0, "... until row {nrows}", "... {nrows} rows in total ..."),
  format = c(NULL, "html", "latex"),
  row.names = FALSE,
  full_width = TRUE,
  position = c("center", "left", "right", "float_left", "float_right"),
  options
)
}
\arguments{
\item{data}{A data frame to display}

\item{top}{An integer. How many of the first rows to show.}

\item{bottom}{Like \code{top}, for for the bottom rows.}

\item{message}{Character string to place between the top and bottom rows.}

\item{format}{Either \code{"html"} or \code{"latex"}.}

\item{row.names}{If \code{TRUE}, will include row.names in table}

\item{full_width}{If \code{FALSE}, table will be condensed}

\item{position}{can be "center", "left", "right", "float_left", or "float_right"}

\item{options}{For compatibility with knitr \verb{render=} chunk option

For Rmd documents, the format will be detected automatically if you don't use the
format argument.`}
}
\description{
Formats tables to show top and bottom rows.
}
