% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset-exercises.R
\name{format_exercises}
\alias{format_exercises}
\title{Utilities for laying out exercises}
\usage{
format_exercises(
  Roster,
  label_format = c("numeric", "alpha", "ALPHA", "roman", "ROMAN"),
  template = "## Exer. {number}\\n\\n **Exercise {number}**\\n\\n{contents}\\n\\n",
  askMC = Znotes::askMC
)
}
\arguments{
\item{Roster}{A data frame with a \code{fname} column
containing the file name of the exercise}

\item{template}{Glue string setting formatting for each exercise}

\item{askMC}{function name for the askMC() typesetting}

\item{show_answers}{Logical: whether to show the answers}
}
\description{
Utilities for laying out exercises
}
